Add backend APIs and Postman collection

+++ Backend/routes/videos.py
@@ -1,10 +1,11 @@
 from flask import Blueprint, request
 import utils.db as db
 from utils.util import createResult
-
+from flask_jwt_extended import jwt_required, get_jwt
 videosRouter = Blueprint("videos", __name__, url_prefix="/video")
 
 @videosRouter.get("/all-videos")
+@jwt_required()
 def get_all_videos():
     courseId = request.args.get("courseId")
 
@@ -18,9 +19,12 @@ def get_all_videos():
     return createResult(None, result)
 
 @videosRouter.post("/add")
+@jwt_required()
 def add_video():
+    claims=get_jwt()
+    if claims.get("role")!="admin":
+          return createResult("Not a Admin",None), 403
     data = request.json
-
     sql = """
     INSERT INTO videos(course_id, title, youtube_url, description)
     VALUES (%s, %s, %s, %s)
@@ -36,7 +40,11 @@ def add_video():
     return createResult(None, result)
 
 @videosRouter.put("/update/<int:videoId>")
+@jwt_required()
 def update_video(videoId):
+    claims=get_jwt()
+    if claims.get("role")!="admin":
+          return createResult("Not a Admin",None), 403
     data = request.json
 
     sql = """
@@ -56,7 +64,12 @@ def update_video(videoId):
     return createResult(None, result)
 
 @videosRouter.delete("/delete/<int:videoId>")
+@jwt_required()
 def delete_video(videoId):
+    claims=get_jwt()
+    if claims.get("role")!="admin":
+          return createResult("Not a Admin",None), 403
     sql = "DELETE FROM videos WHERE video_id = %s"
     result = db.executeQuery(sql, (videoId,))
     return createResult(None, result)
+
